MB-68266: Fix narrowing conversion warnings in cluster_framework
Replace implicit narrowing conversions from size_t to int and
Vbid::id_type with explicit gsl::narrow<>() to satisfy compiler
warnings while preserving intent. Vbucket indices and node assignments
are constrained by their respective counts, making these conversions
safe.

Reviewed-on: https://review.couchbase.org/c/kv_engine/+/244297
Tested-by: Build Bot <build@couchbase.com>

+++ cluster_framework/bucket.cc
@@ -39,14 +39,14 @@ Bucket::Bucket(const Cluster& cluster,
         for (size_t vb = 0; vb < vbuckets; ++vb) {
             vbucketmap[vb].resize(replicas + 1);
             for (size_t n = 0; n < (replicas + 1); ++n) {
-                vbucketmap[vb][n] = (vb + ii++) % nodes;
+                vbucketmap[vb][n] = gsl::narrow<int>((vb + ii++) % nodes);
             }
         }
     } else {
         for (size_t vb = 0; vb < vbuckets; ++vb) {
             vbucketmap[vb].resize(replicas + 1);
             for (size_t n = 0; n < (replicas + 1); ++n) {
-                vbucketmap[vb][n] = ii++ % nodes;
+                vbucketmap[vb][n] = gsl::narrow<int>(ii++ % nodes);
             }
         }
     }
